pyproject.toml
[tool.poetry]
name = "flash51"
version = "0.1.0"
description = ""
authors = ["Chenghsuan Ho <chengxuanjhe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyserial = "^3.5"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
